Simplifying configuration model. Moved test size to plugin extension and CLI configuration file.
