+++ fork-gradle-plugin/src/main/groovy/com/shazam/fork/gradle/ForkExtension.groovy
@@ -58,6 +58,11 @@ class ForkExtension {
     int testOutputTimeout
 
     /**
+     * The size of the tests that will be executed with this run.
+     */
+    String testSize
+
+    /**
      * Indicate that screenshots are allowed when videos are not supported.
      */
     boolean fallbackToScreenshots
