+++ fork-runner/src/main/java/com/shazam/fork/system/EnvironmentConstants.java
@@ -7,5 +7,4 @@ class EnvironmentConstants {
     static final String PARAMETER_POOL_COMPUTED = "fork.computed.";
     static final String PARAMETER_POOL_EACHDEVICE = "fork.eachdevice";
     static final String PARAMETER_EXCLUDED_SERIALS = "fork.excluded.serial";
-    static final String PARAMETER_TEST_SIZE = "fork.test.size";
 }
